requirements.txt
fastapi
uvicorn
jinja2
pydantic[email]
google-generativeai
python-dotenv
flask
gunicorn
